requirements.txt
celery==5.2.3
celery-progress==0.1.2
coverage[toml]==6.3.2
Django==3.2.12
django-auth-ldap==4.0.0
django-autocomplete-light==3.9.4
django-bootstrap-datepicker-plus==4.0.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
# django-ckeditor from git while waiting for a new release (>6.2.0)
git+https://github.com/django-ckeditor/django-ckeditor@50cd4ab#egg=django-ckeditor
django-crispy-forms==1.14.0
django-filter==21.1
django-hijack==3.1.6
django-imagekit==4.1.0
django-import-export==2.7.1
django-redis==5.2.0
django-select2==7.10.0
django-simple-menu==1.2.1
django-storages==1.12.3
django-tables2==2.4.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
# docx-mailmerge from git while waiting for a new release (>0.5.0)
git+https://github.com/Bouke/docx-mailmerge@7024217#egg=docx-mailmerge
environs==9.5.0
factory-boy==3.2.1
fpdf==1.7.2
freezegun==1.2.1
lxml==4.8.0
marshmallow==3.15.0
mkdocs==1.3.0
mkdocs-material==8.2.8
mkdocs-material-extensions==1.0.3
mssql-django==1.1.2
parameterized==0.8.1
paramiko==2.10.3
Pillow==9.0.1
pre-commit==2.17.0
pydantic==1.9.0
PyPDF2==1.26.0
python-dateutil==2.8.2
python-ldap==3.4.0
PyYAML==6.0
requests==2.27.1
sentry-sdk==1.5.8
uk-postcode-utils==1.1
unittest-xml-reporting==3.2.0
weasyprint==52.5  # pinned while dev environments run on Debian 10
xlwt==1.3.0
